Data.CharSet

Description

A CharSet is an efficient representation of a set of Char values designed for fast membership tests.

As an example build isAlpha will create a set of alphabetic characters. We can then use member on the generated set to efficiently test if a given Char represents an alphabetic character.

Designed to be imported qualified:

import Data.CharSet (CharSet)
import qualified Data.CharSet as CharSet

Set type

data CharSet

Stored as a (possibly negated) IntSet and a fast set used for the head byte.

The set of valid (possibly negated) head bytes is stored unboxed as a 32-byte bytestring-based lookup table.

Constructors

CharSet !Bool !ByteSet !IntSet

Operators

(\\) :: CharSet -> CharSet -> CharSet

Query

null :: CharSet -> Bool

O(n) worst case

size :: CharSet -> Int

O(n)

member :: Char -> CharSet -> Bool

notMember :: Char -> CharSet -> Bool

overlaps :: CharSet -> CharSet -> Bool

isSubsetOf :: CharSet -> CharSet -> Bool

isComplemented :: CharSet -> Bool

Construction

build :: (Char -> Bool) -> CharSet

Applies a predicate across the whole range of possible character values to create a set of only those characters which satisfy the predicate.

As an example build isAlpha will generate a CharSet of all alphabetic characters.

empty :: CharSet

singleton :: Char -> CharSet

full :: CharSet

insert :: Char -> CharSet -> CharSet

delete :: Char -> CharSet -> CharSet

complement :: CharSet -> CharSet

range :: Char -> Char -> CharSet

Combine

union :: CharSet -> CharSet -> CharSet

intersection :: CharSet -> CharSet -> CharSet

difference :: CharSet -> CharSet -> CharSet

Filter

filter :: (Char -> Bool) -> CharSet -> CharSet

partition :: (Char -> Bool) -> CharSet -> (CharSet, CharSet)

Map

map :: (Char -> Char) -> CharSet -> CharSet

Fold

fold :: (Char -> b -> b) -> b -> CharSet -> b

Conversion

List

toList :: CharSet -> String

fromList :: String -> CharSet

Ordered list

toAscList :: CharSet -> String

fromAscList :: String -> CharSet

fromDistinctAscList :: String -> CharSet

IntMaps

fromCharSet :: CharSet -> (Bool, IntSet)

toCharSet :: IntSet -> CharSet

Array

toArray :: CharSet -> UArray Char Bool