Data.CharSet.ByteSet

Description

Fast set membership tests for byte values. The set representation is unboxed for efficiency and uses one bit per byte to represent the presence or absence of a byte in the set.

This is a fairly minimal API. You probably want to use CharSet.

Data type

newtype ByteSet

Constructors

ByteSet ByteString

Construction

fromList :: [Word8] -> ByteSet

Lookup

member :: Word8 -> ByteSet -> Bool

Check the set for membership.