Data.CharSet.Unicode.Block

Description

Provides unicode general categories, which are typically connoted by p{InBasicLatin} or p{InIPA_Extensions}. Lookups can be constructed using categories or individual character sets can be used directly.

Unicode General Category

data Block

Constructors

Block

Fields

blockName :: String

blockCharSet :: CharSet

Lookup

blocks :: [Block]

lookupBlock :: String -> Maybe Block

lookupBlockCharSet :: String -> Maybe CharSet

CharSets by Block

basicLatin :: CharSet

latin1Supplement :: CharSet

latinExtendedA :: CharSet

latinExtendedB :: CharSet

ipaExtensions :: CharSet

spacingModifierLetters :: CharSet

combiningDiacriticalMarks :: CharSet

greekAndCoptic :: CharSet

cyrillic :: CharSet

cyrillicSupplementary :: CharSet

armenian :: CharSet

hebrew :: CharSet

arabic :: CharSet

syriac :: CharSet

thaana :: CharSet

devanagari :: CharSet

bengali :: CharSet

gurmukhi :: CharSet

gujarati :: CharSet

oriya :: CharSet

tamil :: CharSet

telugu :: CharSet

kannada :: CharSet

malayalam :: CharSet

sinhala :: CharSet

thai :: CharSet

lao :: CharSet

tibetan :: CharSet

myanmar :: CharSet

georgian :: CharSet

hangulJamo :: CharSet

ethiopic :: CharSet

cherokee :: CharSet

unifiedCanadianAboriginalSyllabics :: CharSet

ogham :: CharSet

runic :: CharSet

tagalog :: CharSet

hanunoo :: CharSet

buhid :: CharSet

tagbanwa :: CharSet

khmer :: CharSet

mongolian :: CharSet

limbu :: CharSet

taiLe :: CharSet

khmerSymbols :: CharSet

phoneticExtensions :: CharSet

latinExtendedAdditional :: CharSet

greekExtended :: CharSet

generalPunctuation :: CharSet

superscriptsAndSubscripts :: CharSet

currencySymbols :: CharSet

combiningDiacriticalMarksForSymbols :: CharSet

letterlikeSymbols :: CharSet

numberForms :: CharSet

arrows :: CharSet

mathematicalOperators :: CharSet

miscellaneousTechnical :: CharSet

controlPictures :: CharSet

opticalCharacterRecognition :: CharSet

enclosedAlphanumerics :: CharSet

boxDrawing :: CharSet

blockElements :: CharSet

geometricShapes :: CharSet

miscellaneousSymbols :: CharSet

dingbats :: CharSet

miscellaneousMathematicalSymbolsA :: CharSet

supplementalArrowsA :: CharSet

braillePatterns :: CharSet

supplementalArrowsB :: CharSet

miscellaneousMathematicalSymbolsB :: CharSet

supplementalMathematicalOperators :: CharSet

miscellaneousSymbolsAndArrows :: CharSet

cjkRadicalsSupplement :: CharSet

kangxiRadicals :: CharSet

ideographicDescriptionCharacters :: CharSet

cjkSymbolsAndPunctuation :: CharSet

hiragana :: CharSet

katakana :: CharSet

bopomofo :: CharSet

hangulCompatibilityJamo :: CharSet

kanbun :: CharSet

bopomofoExtended :: CharSet

katakanaPhoneticExtensions :: CharSet

enclosedCjkLettersAndMonths :: CharSet

cjkCompatibility :: CharSet

cjkUnifiedIdeographsExtensionA :: CharSet

yijingHexagramSymbols :: CharSet

cjkUnifiedIdeographs :: CharSet

yiSyllables :: CharSet

yiRadicals :: CharSet

hangulSyllables :: CharSet

highSurrogates :: CharSet

highPrivateUseSurrogates :: CharSet

lowSurrogates :: CharSet

privateUseArea :: CharSet

cjkCompatibilityIdeographs :: CharSet

alphabeticPresentationForms :: CharSet

arabicPresentationFormsA :: CharSet

variationSelectors :: CharSet

combiningHalfMarks :: CharSet

cjkCompatibilityForms :: CharSet

smallFormVariants :: CharSet

arabicPresentationFormsB :: CharSet

halfwidthAndFullwidthForms :: CharSet

specials :: CharSet