Data.CharSet.Unicode.Category

Description

Provides unicode general categories, which are typically connoted by \p{Ll} or \p{Modifier_Letter}. Lookups can be constructed using categories or individual character sets can be used directly.

A case, _ and - insensitive lookup is provided by lookupCategory and can be used to provide behavior similar to that of Perl or PCRE.

Unicode General Category

data Category

Constructors

Category

Fields

categoryName :: String

categoryAbbreviation :: String

categoryCharSet :: CharSet

categoryDescription :: String

Lookup

categories :: [Category]

lookupCategory :: String -> Maybe Category

lookupCategoryCharSet :: String -> Maybe CharSet

CharSets by Category

Letter

modifierLetter :: CharSet

otherLetter :: CharSet

letter :: CharSet

Letter&

lowercaseLetter :: CharSet

uppercaseLetter :: CharSet

titlecaseLetter :: CharSet

letterAnd :: CharSet

Mark

nonSpacingMark :: CharSet

spacingCombiningMark :: CharSet

enclosingMark :: CharSet

mark :: CharSet

Separator

space :: CharSet

lineSeparator :: CharSet

paragraphSeparator :: CharSet

separator :: CharSet

Symbol

mathSymbol :: CharSet

currencySymbol :: CharSet

modifierSymbol :: CharSet

otherSymbol :: CharSet

symbol :: CharSet

Number

decimalNumber :: CharSet

letterNumber :: CharSet

otherNumber :: CharSet

number :: CharSet

Punctuation

dashPunctuation :: CharSet

openPunctuation :: CharSet

closePunctuation :: CharSet

initialQuote :: CharSet

finalQuote :: CharSet

connectorPunctuation :: CharSet

otherPunctuation :: CharSet

punctuation :: CharSet

Other

control :: CharSet

format :: CharSet

privateUse :: CharSet

surrogate :: CharSet

notAssigned :: CharSet

other :: CharSet